+++ R/pipe.R
@@ -1,3 +1,7 @@
 #' @importFrom magrittr %>%
 #' @export
 magrittr::`%>%`
+
+#' @importFrom htmltools tags
+#' @export
+htmltools::tags