+++ R/text.R
@@ -29,8 +29,8 @@
 #' sigmajs() %>%
 #'   sg_nodes(nodes, id, label, size, color) %>%
 #'   sg_add_nodes(nodes2, delay, id, label, size, color) %>% 
-#'   sg_button("add", "add_nodes") %>% 
-#'   sg_progress(nodes2, delay, text, element = "h3")
+#'   sg_progress(nodes2, delay, text, element = "h3") %>%
+#'   sg_button(c("add_nodes", "progress"), "add") 
 #' 
 #' @export
 sg_progress <- function(sg, data, delay, text, position = "right",