+++ R/utils.R
@@ -88,7 +88,7 @@ globalVariables(c("from", "to", "."))
     "drag_nodes", "relative_size", "add_nodes", 
     "add_edges", "drop_nodes", "drop_edges", 
     "animate", "export_svg", "export_img",
-    "add_nodes_edges")
+    "add_nodes_edges", "progress")
 }